+++ KERN/nke/sctp_nke/SnowLeopard/netinet/sctp_auth.c
@@ -259,7 +259,7 @@ sctp_unpack_auth_chunks(const uint8_t *ptr, uint8_t num_chunks,
  * allocate structure space for a key of length keylen
  */
 sctp_key_t *
-sctp_alloc_key(uint32_t keylen)
+sctp_alloc_key(size_t keylen)
 {
 	sctp_key_t *new_key;
 
@@ -318,7 +318,7 @@ sctp_show_key(sctp_key_t *key, const char *str)
 	}
 }
 
-static uint32_t
+static size_t
 sctp_get_keylen(sctp_key_t *key)
 {
 	if (key != NULL)
@@ -331,7 +331,7 @@ sctp_get_keylen(sctp_key_t *key)
  * generate a new random key of length 'keylen'
  */
 sctp_key_t *
-sctp_generate_random_key(uint32_t keylen)
+sctp_generate_random_key(size_t keylen)
 {
 	sctp_key_t *new_key;
 
@@ -350,7 +350,7 @@ sctp_generate_random_key(uint32_t keylen)
 }
 
 sctp_key_t *
-sctp_set_key(uint8_t *key, uint32_t keylen)
+sctp_set_key(uint8_t *key, size_t keylen)
 {
 	sctp_key_t *new_key;
 
@@ -437,7 +437,7 @@ sctp_compare_key(sctp_key_t *key1, sctp_key_t *key2)
 sctp_key_t *
 sctp_compute_hashkey(sctp_key_t *key1, sctp_key_t *key2, sctp_key_t *shared)
 {
-	uint32_t keylen;
+	size_t keylen;
 	sctp_key_t *new_key;
 	uint8_t *key_ptr;
 
@@ -811,9 +811,9 @@ sctp_serialize_hmaclist(sctp_hmaclist_t *list, uint8_t *ptr)
 }
 
 int
-sctp_verify_hmac_param (struct sctp_auth_hmac_algo *hmacs, uint32_t num_hmacs)
+sctp_verify_hmac_param (struct sctp_auth_hmac_algo *hmacs, size_t num_hmacs)
 {
-	uint32_t i;
+	size_t i;
 	uint16_t hmac_id;
 	uint32_t sha1_supported = 0;
