+++ KERN/nke/sctp_nke/SnowLeopard/netinet/sctp_os_macosx.h
@@ -473,7 +473,7 @@ typedef struct rtentry	sctp_rtentry_t;
 }
 
 struct mbuf *
-sctp_get_mbuf_for_msg(unsigned int space_needed,
+sctp_get_mbuf_for_msg(size_t space_needed,
 		      int want_header, int how, int allonebuf, int type);
 
 /*
